[update] move user.model to user.model.old -> New model for the new mysql connection will be used later
